restructuring
changed the homepage navigation and task manager layout
