+++ src/app/fellowships/[fellowship_id]/page.tsx
@@ -1,8 +1,6 @@
 import Image from 'next/image'
-// Assuming your Footer component is a named export 'Footer'
-import { Footer } from '@/components/footer'
-// Assuming your Header component is a named export 'Header'
-import { Header } from '@/components/header' // <--- ADD THIS LINE FOR THE HEADER
+
+import { Header } from '@/components/header' 
 import { Heart, Calendar } from 'lucide-react'
 import { Button } from '@/components/ui/button'
 import { Card, CardContent } from '@/components/ui/card'
@@ -14,7 +12,7 @@ export default function Home() {
   return (
     <div className="flex min-h-screen flex-col items-center">
      
-      <Header /> {/* <--- ADD THIS LINE FOR THE HEADER */}
+      <Header /> 
 
       <main className="flex-1 w-full">
         {/* Hero Section */}
@@ -286,7 +284,7 @@ export default function Home() {
       </main>
 
    
-      <Footer /> 
+    
     </div>
   )
 }
