--- src/app/fellowships/[fellowship_id]/task-manager/main/tasks/loading.tsx
@@ -1,3 +1,0 @@
-export default function Loading() {
-  return null
-}
