--- src/app/fellowships/[fellowship_id]/task-manager/page.tsx
@@ -1,142 +1,0 @@
-"use client"
-
-import { motion } from "framer-motion"
-import { ArrowRight, CheckCircle, Clock, Users, Zap } from "lucide-react"
-import Link from "next/link"
-import { Button } from "@/components/ui/button"
-import { Card, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
-
-export default function LandingPage() {
-  return (
-    <div className="min-h-screen bg-gray-900 text-white">
-      {/* Hero Section */}
-      <section className="relative overflow-hidden">
-        <div className="absolute inset-0 bg-gradient-to-br from-blue-600/20 to-purple-600/20" />
-        <div className="relative container mx-auto px-4 py-20">
-          <motion.div
-            initial={{ opacity: 0, y: 20 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ duration: 0.8 }}
-            className="text-center max-w-4xl mx-auto"
-          >
-            <h1 className="text-5xl md:text-7xl font-bold mb-6 gradient-text">Taskify</h1>
-            <p className="text-xl md:text-2xl text-gray-300 mb-8 leading-relaxed">
-              The modern task management platform that helps teams stay organized, productive, and focused on what
-              matters most.
-            </p>
-            <div className="flex flex-col sm:flex-row gap-4 justify-center">
-              <Link href="/auth/register">
-                <Button size="lg" className="bg-blue-600 hover:bg-blue-700 text-white px-8 py-3">
-                  Get Started
-                  <ArrowRight className="ml-2 h-5 w-5" />
-                </Button>
-              </Link>
-              <Link href="/auth/login">
-                <Button
-                  variant="outline"
-                  size="lg"
-                  className="border-gray-600 text-gray-300 hover:bg-gray-800 px-8 py-3 bg-transparent"
-                >
-                  Sign In
-                </Button>
-              </Link>
-            </div>
-          </motion.div>
-        </div>
-      </section>
-
-      {/* Features Section */}
-      <section className="py-20 bg-gray-800/50">
-        <div className="container mx-auto px-4">
-          <motion.div
-            initial={{ opacity: 0, y: 20 }}
-            whileInView={{ opacity: 1, y: 0 }}
-            transition={{ duration: 0.8 }}
-            className="text-center mb-16"
-          >
-            <h2 className="text-4xl font-bold mb-4">Everything you need to manage tasks</h2>
-            <p className="text-xl text-gray-400 max-w-2xl mx-auto">
-              From simple to-dos to complex project management, Taskify adapts to your workflow.
-            </p>
-          </motion.div>
-
-          <div className="grid md:grid-cols-2 lg:grid-cols-4 gap-8">
-            {features.map((feature, index) => (
-              <motion.div
-                key={feature.title}
-                initial={{ opacity: 0, y: 20 }}
-                whileInView={{ opacity: 1, y: 0 }}
-                transition={{ duration: 0.8, delay: index * 0.1 }}
-              >
-                <Card className="bg-gray-800 border-gray-700 hover:bg-gray-750 transition-colors">
-                  <CardHeader>
-                    <div className="w-12 h-12 bg-blue-600/20 rounded-lg flex items-center justify-center mb-4">
-                      <feature.icon className="h-6 w-6 text-blue-400" />
-                    </div>
-                    <CardTitle className="text-white">{feature.title}</CardTitle>
-                    <CardDescription className="text-gray-400">{feature.description}</CardDescription>
-                  </CardHeader>
-                </Card>
-              </motion.div>
-            ))}
-          </div>
-        </div>
-      </section>
-
-      {/* CTA Section */}
-      <section className="py-20">
-        <div className="container mx-auto px-4">
-          <motion.div
-            initial={{ opacity: 0, y: 20 }}
-            whileInView={{ opacity: 1, y: 0 }}
-            transition={{ duration: 0.8 }}
-            className="text-center max-w-3xl mx-auto"
-          >
-            <h2 className="text-4xl font-bold mb-6">Ready to get organized?</h2>
-            <p className="text-xl text-gray-400 mb-8">
-              Join thousands of teams already using Taskify to streamline their workflow.
-            </p>
-            <Link href="/fellowships/fellowship1/task-manager/main/dashboard">
-              <Button size="lg" className="bg-blue-600 hover:bg-blue-700 text-white px-8 py-3">
-                Start Managing Tasks
-                <ArrowRight className="ml-2 h-5 w-5" />
-              </Button>
-            </Link>
-          </motion.div>
-        </div>
-      </section>
-
-      {/* Footer */}
-      <footer className="border-t border-gray-800 py-8">
-        <div className="container mx-auto px-4">
-          <div className="text-center text-gray-400">
-            <p>&copy; 2024 Taskify. Built with Next.js and Tailwind CSS.</p>
-          </div>
-        </div>
-      </footer>
-    </div>
-  )
-}
-
-const features = [
-  {
-    title: "Task Management",
-    description: "Create, organize, and track tasks with powerful filtering and search capabilities.",
-    icon: CheckCircle,
-  },
-  {
-    title: "Team Collaboration",
-    description: "Assign tasks, track progress, and collaborate seamlessly with your team.",
-    icon: Users,
-  },
-  {
-    title: "Time Tracking",
-    description: "Monitor time spent on tasks and projects with built-in time tracking tools.",
-    icon: Clock,
-  },
-  {
-    title: "Automation",
-    description: "Set up recurring tasks and automated workflows to save time and effort.",
-    icon: Zap,
-  },
-]
