+++ src/app/homepage/layout.tsx
@@ -1,0 +1,13 @@
+
+import { HomeHeader } from '@/components/homeHeader'
+import { Footer } from '@/components/footer'
+
+export default function HomeLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <>
+      <HomeHeader />
+      {children}
+      <Footer />
+    </>
+  )
+}
