+++ src/app/layout.tsx
@@ -1,11 +1,11 @@
 
 import '@/styles/globals.css'
+import './globals.css'
 import type { Metadata } from 'next'
-import { Inter } from 'next/font/google'
+import { Inter, Geist } from 'next/font/google'
 import { Toaster } from '@/components/ui/toaster'
 import { AOSProvider } from '@/components/aos-provider'
 import { TRPCReactProvider } from '@/trpc/react'
-import { Geist } from 'next/font/google'
 
 // Load fonts
 const geist = Geist({
@@ -32,7 +32,6 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
       <body className={inter.className}>
         <AOSProvider>
           <TRPCReactProvider>
-           
             <main>{children}</main>
             <Toaster />
           </TRPCReactProvider>
